[Core] Improve base command to allow reuse of internal buffer
The internal buffer size is now set to 2+256 bytes:
 * This allows to handle commands with large number of parameter
 * Removes constraint to provide mutable buffer for write fifo
Overall this saved more than 1k on a typical final application size

New available command: buffer() and buffer_mut gives direct access to
the internal buffer: this remove the need to allocate buffer when read/writing the FIFO

Status: methods related to the command status have been move to the enum directly
to allow other manipulating it directly when the full 8b value is not directly available (like inside the CmdBuffer).

PatchRam:
 * Write command have been update to re-use the internal buffer
 * Add API to check the PatchRam has been set correctly

Function to read/write the fifo where split in two: one where the internal
buffer is used and one where an external buffer is provided:
this leaves maximum flexibility for the final user.